requirements.txt
Jinija==2.9.5
pyconfluence==1.1.3
py-trello==0.8.0